+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/Player/EffectCrouchMode.lua
@@ -10,6 +10,7 @@ end
 
 function EffectCrouchMode:OnStart()
     ChaosEffectBase:OnStart()
+    ChaosVehicle.ExitVehicle(getPlayer())
     Events.OnPlayerUpdate.Add(forceSneak)
 end